# docs/swagger.yaml
basePath: /
definitions:
  models.MonitoringRequest:
    properties:
      admin:
        type: string
      adminId:
        type: integer
      creationDate:
        type: string
      creator:
        type: string
      endingDate:
        type: string
      formationDate:
        type: string
      requestId:
        type: integer
      status:
        type: string
      userId:
        type: integer
    type: object
  models.NewStatus:
    properties:
      status:
        type: string
    type: object
...
host: localhost:3001
info:
  contact: {}
  description: App for serving threats monitoring requests
  title: ThreatMonitoringApp
  version: "1.0"
paths:
...
  /api/monitoring-request-threats/threats/{threatId}:
    delete:
      consumes:
      - application/json
      description: Deletes a threat from a request based on the user ID and threat
        ID
      parameters:
      - description: Threat ID
        in: path
        name: threatId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema: {}
      summary: Delete threat from request
      tags:
      - MonitoringRequests
  /api/monitoring-requests:
    delete:
      consumes:
      - application/json
      description: Deletes a monitoring request for the given user ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema: {}
      summary: Delete monitoring request by user ID
      tags:
      - MonitoringRequests
    get:
      consumes:
      - application/json
      description: Retrieves a list of monitoring requests based on the provided parameters
      parameters:
      - description: Monitoring request status
        in: query
        name: status
        type: string
      - description: Start date in the format '2006-01-02T15:04:05Z'
        in: query
        name: start_date
        type: string
      - description: End date in the format '2006-01-02T15:04:05Z'
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.MonitoringRequest'
            type: array
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get list of monitoring requests
      tags:
      - MonitoringRequests
  /api/monitoring-requests/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a monitoring request with the given ID
      parameters:
      - description: Monitoring Request ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema: {}
      summary: Get monitoring request by ID
      tags:
      - MonitoringRequests
  /api/monitoring-requests/client:
    put:
      consumes:
      - application/json
      description: Updates the status of a monitoring request by client on formated
      parameters:
      - description: New status of the monitoring request
        in: body
        name: newStatus
        required: true
        schema:
          $ref: '#/definitions/models.NewStatus'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema: {}
      summary: Update monitoring request status by client
      tags:
      - MonitoringRequests
...
  /api/threats/request/{threatId}:
    post:
      consumes:
      - application/json
      description: Adds a threat to a monitoring request
      parameters:
      - description: Threat ID
        in: path
        name: threatId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema: {}
      summary: Add threat to request
      tags:
      - Threats
  /monitoring-requests/admin/{requestId}:
    put:
      consumes:
      - application/json
      description: Updates the status of a monitoring request with the given ID on
        "accepted"/"closed"/"canceled"
      parameters:
      - description: Request ID
        in: path
        name: requestId
        required: true
        type: integer
      - description: New request status
        in: body
        name: newRequestStatus
        required: true
        schema:
          $ref: '#/definitions/models.NewStatus'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema: {}
      summary: Update monitoring request status by ID
      tags:
      - MonitoringRequests
schemes:
- http
swagger: "2.0"